+++ packages/legend-application-query/src/index.ts
@@ -49,7 +49,3 @@ export {
   StandardQueryBuilderMode,
   QueryBuilderState,
 } from './stores/QueryBuilderState.js';
-export { QueryBuilderExplorerTreeRootNodeData } from './stores/explorer/QueryBuilderExplorerState.js';
-export { FETCH_STRUCTURE_MODE } from './stores/fetch-structure/QueryBuilderFetchStructureState.js';
-export { COLUMN_SORT_TYPE } from './stores/fetch-structure/projection/QueryResultSetModifierState.js';
-export { QueryBuilderSimpleProjectionColumnState } from './stores/fetch-structure/projection/QueryBuilderProjectionState.js';
