+++ packages/legend-application-query/src/stores/fetch-structure/projection/aggregation/operators/QueryBuilderAggregateOperator_Average.ts
@@ -29,7 +29,7 @@ import { QueryBuilderAggregateOperator } from '../QueryBuilderAggregateOperator.
 import {
   type QueryBuilderProjectionColumnState,
   QueryBuilderSimpleProjectionColumnState,
-} from '../../QueryBuilderProjectionState.js';
+} from '../../QueryBuilderProjectionColumnState.js';
 import {
   buildAggregateColumnState,
   buildAggregateExpression,
